method for recovering original source and receiver primary keys

+++ noisemodelling-propagation/src/main/java/org/noise_planet/noisemodelling/propagation/IComputeRaysOut.java
@@ -11,13 +11,13 @@ public interface IComputeRaysOut {
      * @param propagationPath Propagation path result
      * @return Optional minimal energetic contribution per frequency band or empty array
      */
-    double[] addPropagationPaths(int sourceId, double sourceLi, int receiverId, List<PropagationPath> propagationPath);
+    double[] addPropagationPaths(long sourceId, double sourceLi, long receiverId, List<PropagationPath> propagationPath);
 
     /**
      * No more propagation paths will be pushed for this receiver identifier
      * @param receiverId
      */
-    void finalizeReceiver(int receiverId);
+    void finalizeReceiver(long receiverId);
     /**
      * If the implementation does not support thread concurrency, this method is called to return an instance
      * @param receiverStart
